feat: 게임이 끝나지 않았으면 삭제하지 못하게 구현

+++ src/main/java/chess/service/ChessService.java
@@ -98,6 +98,11 @@ public class ChessService {
 
     public void deleteRoom(int gameId, String inputPassword) {
         gameDao.checkCanDelete(gameId, inputPassword);
+        ChessGame chessGame = generateChessGame(gameId);
+        if (!chessGame.isEnd()) {
+            throw new IllegalArgumentException("게임이 아직 안끝났습니다!");
+        }
+
         boardDao.deletePieces(gameId);
         gameDao.deleteGame(gameId);
     }
